refactor(wingman): move threaded_execution to threading_utils
Wingman.threaded_execution is now a one-line wrapper around the free
function in services/threading_utils.py. The wrapper is kept because
WingmanSkillManager binds it onto every skill instance.
